requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.3
beautifulsoup4==4.7.1
black==18.9b0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cssselect==1.0.3
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.9
en-core-web-sm==2.0.0
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
isort==4.3.15
jieba3k==0.35.1
lazy-object-proxy==1.3.1
lxml==4.3.2
mccabe==0.6.1
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
newspaper3k==0.2.8
nltk==3.4
numpy==1.16.2
Pillow==5.4.1
plac==0.9.6
preshed==2.0.1
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.0
PyYAML==3.13
regex==2018.1.10
requests==2.21.0
requests-file==1.4.3
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.8
spacy==2.0.18
thinc==6.12.1
tinysegmenter==0.3
tldextract==2.2.1
toml==0.10.0
toolz==0.9.0
tqdm==4.31.1
typed-ast==1.3.1
ujson==1.35
urllib3==1.24.1
wrapt==1.10.11
